Definetly getting rid of getStaticParams
