+++ utils/supabase/middleware.ts
@@ -58,3 +58,7 @@ export async function updateSession(request: NextRequest) {
 
   return response;
 }
+
+// export const config = {
+//   matcher: ["/admin/:path*", "/login", "/logout"],
+// };
