+++ utils/supabase/server.ts
@@ -1,40 +1,3 @@
-// import { createServerClient } from "@supabase/ssr";
-// import { cookies } from "next/headers";
-
-// export const createClient = () => {
-//   const cookieStore = cookies();
-
-//   return createServerClient(
-//     process.env.NEXT_PUBLIC_SUPABASE_URL,
-//     process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY,
-//     {
-//       cookies: {
-//         get(name) {
-//           return cookieStore.get(name)?.value;
-//         },
-//         set(name, value, options) {
-//           try {
-//             cookieStore.set({ name, value, ...options });
-//           } catch (error) {
-//             // The `set` method was called from a Server Component.
-//             // This can be ignored if you have middleware refreshing
-//             // user sessions.
-//           }
-//         },
-//         remove(name, options) {
-//           try {
-//             cookieStore.set({ name, value: "", ...options });
-//           } catch (error) {
-//             // The `delete` method was called from a Server Component.
-//             // This can be ignored if you have middleware refreshing
-//             // user sessions.
-//           }
-//         },
-//       },
-//     }
-//   );
-// };
-
 import { createServerClient, type CookieOptions } from "@supabase/ssr";
 import { cookies } from "next/headers";
 
@@ -67,13 +30,3 @@ export function createClient() {
     }
   );
 }
-
-//utils/supabase/server.js
-// import { createClient as createSupabaseClient } from "@supabase/supabase-js";
-
-// export const createClient = () => {
-//   return createSupabaseClient(
-//     process.env.NEXT_PUBLIC_SUPABASE_URL!,
-//     process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
-//   );
-// };
